Workaround for a float -> double conversion bug
Joshua Braegger reported a bug caused by a float to double conversion.  The
bug is exemplified by the following Groovy code:

    float maxWidth = 440
    int width = 2000
    float modifier = maxWidth / width
    println( (double) modifier) // prints 0.219999 ... should print 0.22

This affected ImageTool.thumbnailSpecial, which needs to keep a modifier
value as either a float or a double and convert to the other, since the
JAI APIs sometimes receive one and sometimes the other.

On my tests, the bug does not appear when keeping the value as a double and
converting to a float, so I've changed it to that implementation.